# docs/specs/05-http-api-openapi.yaml
openapi: 3.1.0
info:
  title: IT Service Ticket System Core API
  version: 1.0.0
  description: |
    非 AI 核心 HTTP 契约。AI/RAG 接口由 AI-005 单独定义；字段和枚举引用 DM-002/DM-004，动作引用 SM-*，错误与重试引用 RD-*。
servers:
  - url: /api/v1
security:
  - ssoSession: []
    requestId: []
...
paths:
  /consultations/{id}/messages:
    post:
      tags: [Consultations]
      operationId: sendHumanConsultationMessage
      summary: 发送人工咨询消息
      parameters: [{ $ref: '#/components/parameters/SessionId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/RequestId' }]
      requestBody: { $ref: '#/components/requestBodies/Content' }
      responses: { '200': { $ref: '#/components/responses/Envelope' }, default: { $ref: '#/components/responses/Error' } }
    get:
      tags: [Consultations]
      operationId: listConsultationMessages
      parameters: [{ $ref: '#/components/parameters/SessionId' }, { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/PageSize' }, { $ref: '#/components/parameters/RequestId' }]
      responses: { '200': { $ref: '#/components/responses/Paged' }, default: { $ref: '#/components/responses/Error' } }
  /consultations/{id}/resolution:
    post:
      tags: [Consultations]
      operationId: resolveConsultation
      summary: 提交咨询解决结论
      parameters: [{ $ref: '#/components/parameters/SessionId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/RequestId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ConsultationResolution' } } } }
      responses: { '200': { $ref: '#/components/responses/Envelope' }, default: { $ref: '#/components/responses/Error' } }
  /consultations/{id}/reopen:
    post:
      tags: [Consultations]
      operationId: reopenConsultation
      summary: 24 小时窗口内恢复咨询
      parameters: [{ $ref: '#/components/parameters/SessionId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/RequestId' }]
      requestBody: { $ref: '#/components/requestBodies/Reason' }
      responses: { '200': { $ref: '#/components/responses/Envelope' }, default: { $ref: '#/components/responses/Error' } }
  /consultations/{id}/ticket-draft:
    get:
      tags: [Consultations]
      operationId: getTicketDraftFromConsultation
      summary: 获取咨询转单预填
      parameters: [{ $ref: '#/components/parameters/SessionId' }]
      responses: { '200': { $ref: '#/components/responses/Envelope' }, default: { $ref: '#/components/responses/Error' } }
...
  /tickets:
    post:
      tags: [Tickets]
      operationId: createTicket
      summary: 提交正式工单
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/RequestId' }]
      requestBody: { $ref: '#/components/requestBodies/TicketCreate' }
      responses: { '201': { $ref: '#/components/responses/Envelope' }, default: { $ref: '#/components/responses/Error' } }
    get:
      tags: [Tickets]
      operationId: listTickets
      summary: 按服务端权限查询工单
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { name: status, in: query, schema: { type: string }, description: 仅接受 DM-002 的 TicketStatus }
        - { name: ownerScope, in: query, schema: { type: string, enum: [MINE, CURRENT, HISTORY, ALL] }, default: MINE }
      responses: { '200': { $ref: '#/components/responses/Paged' }, default: { $ref: '#/components/responses/Error' } }
  /tickets/{id}:
    get:
      tags: [Tickets]
      operationId: getTicket
      summary: 获取按权限裁剪的工单详情
      parameters: [{ $ref: '#/components/parameters/TicketId' }]
      responses: { '200': { $ref: '#/components/responses/Envelope' }, default: { $ref: '#/components/responses/Error' } }
...
  /consultations/{id}:
    get:
      tags: [Consultations]
      operationId: getConsultation
      parameters: [{ $ref: '#/components/parameters/SessionId' }, { $ref: '#/components/parameters/RequestId' }]
      responses: { '200': { $ref: '#/components/responses/Consultation' }, '401': { $ref: '#/components/responses/Error' }, '403': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' } }
  /consultations/{id}/close:
    post:
      tags: [Consultations]
      operationId: closeConsultation
      parameters: [{ $ref: '#/components/parameters/SessionId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/RequestId' }]
      requestBody: { $ref: '#/components/requestBodies/Reason' }
      responses: { '200': { $ref: '#/components/responses/Consultation' }, '409': { $ref: '#/components/responses/Error' }, default: { $ref: '#/components/responses/Error' } }
...
  /consultations:
    get:
      tags: [Consultations]
      operationId: listConsultations
      parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/PageSize' }, { $ref: '#/components/parameters/RequestId' }]
      responses: { '200': { $ref: '#/components/responses/Paged' }, default: { $ref: '#/components/responses/Error' } }
...
components:
  securitySchemes:
    ssoSession: { type: http, scheme: bearer, bearerFormat: SSO-session }
    requestId: { type: apiKey, in: header, name: X-Request-Id }
  parameters:
    SessionId: { name: id, in: path, required: true, schema: { type: string, maxLength: 64 } }
    TicketId: { name: id, in: path, required: true, schema: { type: string, maxLength: 64 } }
...
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1, maxLength: 128 } }
    RequestId: { name: X-Request-Id, in: header, required: true, schema: { type: string, minLength: 1, maxLength: 128 } }
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
  requestBodies:
    Content: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Content' } } } }
    Reason: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Reason' } } } }
...
    TicketCreate: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TicketCreate' } } } }
...
  responses:
    Envelope: { description: Success, content: { application/json: { schema: { $ref: '#/components/schemas/Envelope' } } } }
    Paged: { description: Success, content: { application/json: { schema: { $ref: '#/components/schemas/PagedEnvelope' } } } }
    Error: { description: Error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Consultation: { description: Consultation, content: { application/json: { schema: { $ref: '#/components/schemas/ConsultationEnvelope' } } } }
...
  schemas:
    Envelope: { type: object, required: [code, message, request_id, data], properties: { code: { type: string }, message: { type: string }, request_id: { type: string }, data: { oneOf: [{ $ref: '#/components/schemas/ConsultationProjection' }, { $ref: '#/components/schemas/TicketProjection' }, { $ref: '#/components/schemas/MessageProjection' }, { $ref: '#/components/schemas/AttachmentProjection' }, { $ref: '#/components/schemas/ActionProjection' }] } }, additionalProperties: false }
    PagedEnvelope: { type: object, required: [code, message, request_id, data], properties: { code: { type: string }, message: { type: string }, request_id: { type: string }, data: { type: object, required: [items, page, pageSize, total], properties: { items: { type: array, items: { oneOf: [{ $ref: '#/components/schemas/TicketProjection' }, { $ref: '#/components/schemas/ConsultationProjection' }, { $ref: '#/components/schemas/MessageProjection' }, { $ref: '#/components/schemas/NotificationProjection' }, { $ref: '#/components/schemas/ExceptionProjection' }, { $ref: '#/components/schemas/UserProjection' }, { $ref: '#/components/schemas/TeamProjection' }, { $ref: '#/components/schemas/CategoryProjection' }, { $ref: '#/components/schemas/TransitionProjection' }, { $ref: '#/components/schemas/KnowledgeVersionProjection' }, { $ref: '#/components/schemas/ConfigProjection' }, { $ref: '#/components/schemas/CaseCandidateProjection' }] } }, page: { type: integer }, pageSize: { type: integer }, total: { type: integer } }, additionalProperties: false } }, additionalProperties: false }
    ErrorEnvelope: { type: object, required: [code, message, request_id], properties: { code: { type: string }, message: { type: string }, request_id: { type: string }, errors: { type: array, items: { type: object, required: [field, reason], properties: { field: { type: string }, reason: { type: string }, message: { type: string } }, additionalProperties: false } } }, additionalProperties: false }
    Content: { type: object, required: [client_message_id], additionalProperties: false, anyOf: [{ required: [content] }, { required: [attachment_ids] }], properties: { content: { type: string, minLength: 1, maxLength: 12000 }, client_message_id: { type: string, minLength: 1, maxLength: 64 }, attachment_ids: { type: array, minItems: 1, items: { type: string, maxLength: 64 }, maxItems: 10 } } }
    Reason: { type: object, required: [reason], additionalProperties: false, properties: { reason: { type: string, minLength: 1, maxLength: 2000 } } }
...
    ConsultationResolution: { type: object, required: [conclusion], additionalProperties: false, properties: { conclusion: { type: string, minLength: 1, maxLength: 12000 } } }
    TicketCreate: { type: object, required: [ticket_nature, category_id, title, description, impact_description, urgency_description], additionalProperties: false, properties: { ticket_nature: { type: string, enum: [INCIDENT, SERVICE_REQUEST] }, category_id: { type: string, minLength: 1, maxLength: 64 }, title: { type: string, minLength: 1, maxLength: 100 }, description: { type: string, minLength: 10, maxLength: 5000 }, impact_description: { type: string, minLength: 1, maxLength: 2000 }, urgency_description: { type: string, minLength: 1, maxLength: 2000 }, location: { type: string, maxLength: 255 }, contact: { type: string, maxLength: 255 }, asset_id: { type: string, maxLength: 64 }, source_session_id: { type: string, maxLength: 32 }, field_values: { type: object } } }
...
    UserProjection: { type: object, required: [user_id, display_name, enabled], additionalProperties: false, properties: { user_id: { type: string }, display_name: { type: string }, enabled: { type: boolean } } }
    TeamProjection: { type: object, required: [team_id, name, enabled], additionalProperties: false, properties: { team_id: { type: string }, name: { type: string }, enabled: { type: boolean } } }
    CategoryProjection: { type: object, required: [category_id, name, level, enabled], additionalProperties: false, properties: { category_id: { type: string }, name: { type: string }, level: { type: integer }, enabled: { type: boolean } } }
    TransitionProjection: { type: object, required: [transition_id, event_code, occurred_at], additionalProperties: false, properties: { transition_id: { type: string }, event_code: { type: string }, occurred_at: { type: string, format: date-time } } }
    KnowledgeVersionProjection: { type: object, required: [version_id, version_no], additionalProperties: false, properties: { version_id: { type: string }, version_no: { type: integer }, published_at: { type: [string, 'null'], format: date-time } } }
    ConfigProjection: { type: object, required: [config_type, version], additionalProperties: false, properties: { config_type: { type: string }, version: { type: integer }, payload: { type: object } } }
    CaseCandidateProjection: { type: object, required: [case_id, source_type, masking_status, status], additionalProperties: false, properties: { case_id: { type: string, maxLength: 64 }, source_type: { type: string, enum: [CONSULTATION, TICKET] }, masking_status: { type: string, enum: [NOT_STARTED, IN_PROGRESS, PASSED, FAILED] }, status: { type: string, enum: [PENDING_MASKING, PENDING_REVIEW, REJECTED, ACCEPTED, DISCARDED] } } }
    ConsultationEnvelope: { type: object, required: [code, message, request_id, data], properties: { code: { type: string }, message: { type: string }, request_id: { type: string }, data: { $ref: '#/components/schemas/ConsultationProjection' } }, additionalProperties: false }
...
    ConsultationProjection: { type: object, required: [session_id, status, creator_id], properties: { session_id: { type: string, maxLength: 32 }, status: { type: string }, creator_id: { type: string, maxLength: 64, description: '咨询申请人账号，用于工程师识别申请人' }, current_engineer_id: { type: [string, 'null'], maxLength: 64 }, converted_ticket_id: { type: [string, 'null'], maxLength: 32 } }, additionalProperties: false }
    TicketProjection: { type: object, required: [ticket_id, status, priority, version], properties: { ticket_id: { type: string, maxLength: 32 }, status: { type: string }, priority: { type: string, enum: [HIGH, MEDIUM, LOW] }, assignee_id: { type: [string, 'null'], maxLength: 64 }, version: { type: integer, minimum: 0 } }, additionalProperties: false }
    MessageProjection:
      type: object
      required: [message_id, sent_at]
      additionalProperties: false
      properties:
        message_id: { type: string, maxLength: 64 }
        sender_type: { type: string, enum: [EMPLOYEE, ENGINEER, AI, SYSTEM] }
        content: { type: string }
        citations:
          type: [array, 'null']
          description: AI-003 KnowledgeCitation；仅包含经校验的已发布知识引用
          items: { $ref: '#/components/schemas/KnowledgeCitation' }
        interaction_id: { type: [string, 'null'], maxLength: 64, description: AI 反馈关联 ID；旧消息可缺失 }
        reply_type: { type: [string, 'null'], enum: [ANSWER, CLARIFY, REFUSE, null] }
        refusal_reason: { type: [string, 'null'], enum: [NO_RELIABLE_KNOWLEDGE, LOW_CONFIDENCE, CONFLICTING_KNOWLEDGE, HIGH_RISK_TOPIC, MODEL_UNAVAILABLE, POLICY_BLOCKED, OFF_TOPIC, null] }
        general_answer: { type: [boolean, 'null'], description: 经校验的无知识库引用通用回答；旧消息不推断 }
        sent_at: { type: string, format: date-time }
        withdrawn_at: { type: [string, 'null'], format: date-time }
    KnowledgeCitation:
      type: object
      additionalProperties: false
      required: [articleId, versionId, title, score, snippet]
      properties:
        articleId: { type: string, minLength: 1, maxLength: 64 }
        versionId: { type: string, minLength: 1, maxLength: 64 }
        title: { type: string, minLength: 1, maxLength: 200 }
        score: { type: number, minimum: 0, maximum: 1 }
        snippet: { type: string, minLength: 1, maxLength: 1000 }
    AttachmentProjection: { type: object, required: [attachment_id, scan_status], properties: { attachment_id: { type: string, maxLength: 64 }, scan_status: { type: string, enum: [PENDING, PASSED, REJECTED, ERROR] }, download_url: { type: [string, 'null'], format: uri } }, additionalProperties: false }
    ActionProjection: { type: object, required: [operation_id], properties: { operation_id: { type: string }, accepted: { type: boolean } }, additionalProperties: false }
    NotificationProjection: { type: object, required: [notification_id, status], properties: { notification_id: { type: string }, status: { type: string } }, additionalProperties: false }
    ExceptionProjection: { type: object, required: [exception_id, status], properties: { exception_id: { type: string }, status: { type: string }, object_id: { type: string } }, additionalProperties: false }
...